Update and move membership test method, add some tests
Updates the membership test method according to answers from Marc Uetz, not working yet alas
Adds a test that I think the member_of method should comply with.
